+++ addOns/automation/src/main/java/org/zaproxy/addon/automation/ExtensionAutomation.java
@@ -59,7 +59,7 @@ import org.zaproxy.addon.automation.jobs.SpiderJob;
 import org.zaproxy.zap.ZAP;
 import org.zaproxy.zap.ZAP.ProcessType;
 import org.zaproxy.zap.eventBus.Event;
-import org.zaproxy.zap.extension.stats.ExtensionStats;
+import org.zaproxy.zap.utils.Stats;
 
 public class ExtensionAutomation extends ExtensionAdaptor implements CommandLineListener {
 
@@ -70,6 +70,13 @@ public class ExtensionAutomation extends ExtensionAdaptor implements CommandLine
     public static final String PREFIX = "automation";
 
     public static final String RESOURCES_DIR = "/org/zaproxy/addon/automation/resources/";
+
+    protected static final String PLANS_RUN_STATS = "stats.auto.plans.run";
+    protected static final String TOTAL_JOBS_RUN_STATS = "stats.auto.jobs.run";
+    protected static final String JOBS_RUN_STATS_PREFIX = "stats.auto.job.";
+    protected static final String JOBS_RUN_STATS_POSTFIX = ".run";
+    protected static final String ERROR_COUNT_STATS = "stats.auto.errors";
+    protected static final String WARNING_COUNT_STATS = "stats.auto.warnings";
 
     public static final ImageIcon ICON =
             new ImageIcon(ExtensionAutomation.class.getResource(RESOURCES_DIR + "robot.png"));
@@ -216,20 +223,18 @@ public class ExtensionAutomation extends ExtensionAdaptor implements CommandLine
         }
     }
 
-    private void clearStats() {
-        // Clear stats for any stats tests
-        ExtensionStats extStats =
-                Control.getSingleton().getExtensionLoader().getExtension(ExtensionStats.class);
-        if (extStats != null && extStats.getInMemoryStats() != null) {
-            extStats.getInMemoryStats().allCleared();
-        }
+    private void setPlanFinished(AutomationPlan plan) {
+        plan.setFinished(new Date());
+        AutomationEventPublisher.publishEvent(
+                AutomationEventPublisher.PLAN_FINISHED, plan, plan.getProgress().toMap());
+        Stats.incCounter(ERROR_COUNT_STATS, plan.getProgress().getErrors().size());
+        Stats.incCounter(WARNING_COUNT_STATS, plan.getProgress().getWarnings().size());
     }
 
     public AutomationProgress runPlan(AutomationPlan plan, boolean resetProgress) {
         if (resetProgress) {
             plan.resetProgress();
         }
-        clearStats();
 
         AutomationProgress progress = plan.getProgress();
         AutomationEnvironment env = plan.getEnv();
@@ -242,12 +247,11 @@ public class ExtensionAutomation extends ExtensionAdaptor implements CommandLine
 
         AutomationEventPublisher.publishEvent(
                 AutomationEventPublisher.PLAN_ENV_CREATED, plan, null);
+        Stats.incCounter(PLANS_RUN_STATS);
 
         if (progress.hasErrors() || env.isTimeToQuit()) {
             // If the environment reports an error then no point in continuing
-            AutomationEventPublisher.publishEvent(
-                    AutomationEventPublisher.PLAN_FINISHED, plan, plan.getProgress().toMap());
-            plan.setFinished(new Date());
+            setPlanFinished(plan);
             return progress;
         }
 
@@ -259,6 +263,8 @@ public class ExtensionAutomation extends ExtensionAdaptor implements CommandLine
             job.setStatus(AutomationJob.Status.RUNNING);
             AutomationEventPublisher.publishEvent(AutomationEventPublisher.JOB_STARTED, job, null);
             job.runJob(env, progress);
+            Stats.incCounter(TOTAL_JOBS_RUN_STATS);
+            Stats.incCounter(JOBS_RUN_STATS_PREFIX + job.getType() + JOBS_RUN_STATS_POSTFIX);
             job.logTestsToProgress(progress);
             job.setStatus(AutomationJob.Status.COMPLETED);
             AutomationEventPublisher.publishEvent(
@@ -271,20 +277,12 @@ public class ExtensionAutomation extends ExtensionAdaptor implements CommandLine
                 break;
             }
         }
-        plan.setFinished(new Date());
-
-        AutomationEventPublisher.publishEvent(
-                AutomationEventPublisher.PLAN_FINISHED, plan, plan.getProgress().toMap());
+        setPlanFinished(plan);
         return progress;
     }
 
     public void runPlanAsync(AutomationPlan plan) {
-        new Thread(
-                        () -> {
-                            this.runPlan(plan, true);
-                        },
-                        "ZAP-Automation")
-                .start();
+        new Thread(() -> this.runPlan(plan, true), "ZAP-Automation").start();
     }
 
     public AutomationPlan loadPlan(File f) throws IOException {